update hoe files, move gga4r.rb out of subdirectory
