+++ Manifest.txt
@@ -1,9 +1,6 @@
 Rakefile
-README.txt
-CHANGELOG.txt
+README.rdoc
 Manifest.txt
-setup.rb
-lib/gga4r/gga4r_main.rb
 lib/gga4r.rb
 test/test_helper.rb
 test/gga4r_test.rb
